requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==23.2.0
black==24.4.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
confluent-kafka==2.3.0
exceptiongroup==1.2.1
Faker==25.0.1
h11==0.14.0
idna==3.7
minio==7.2.7
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
pycparser==2.22
pycryptodome==3.20.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
requests==2.31.0
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
wsproto==1.2.0
